+++ packages/visual-editor/src/components/pageSections/ProductSection/ProductCard.tsx
@@ -399,12 +399,10 @@ const ProductCardComponent: PuckComponent<ProductCardProps> = (props) => {
         <div className={variant === "classic" ? "px-8 pt-8" : ""}>
           <div
             className={themeManagerCn(
-              imageConstrain === "fixed" ? "w-fit mx-auto" : "w-full",
-              imageStyles?.aspectRatio && "md:aspect-[var(--aspect-ratio)]"
+              imageConstrain === "fixed" ? "w-fit mx-auto" : "w-full"
             )}
             style={
               {
-                "--aspect-ratio": imageStyles?.aspectRatio,
                 width:
                   imageConstrain === "fixed" && imageStyles?.width
                     ? `${imageStyles.width}px`